normal operation

+++ widgets/source/Storage/DAO/Storage_DAO_Sqlite.cpp
@@ -9,6 +9,7 @@
 #include <QScreen>
 #include <QImage>
 #include <QLabel>
+#include <QSqlRecord>
 /*
  * In current softwares, there are two place to record the operation data:
  * (1) The .csv file recorded the patient therapy time, located in user specialed path
@@ -236,6 +237,8 @@ void DAO_Sqlite::appendARow_TR(const QString & tableName,
 }
 
 void DAO_Sqlite::appendARow_Patient(const std::map<unsigned int,std::pair<QString,QString>> & patientInfos){
+    QStringList list;
+    bool flag = true;
     QSqlQuery query(*this->theDataBase);
     const std::map<unsigned int,OnePatientPattern> * patientInfoPatten = ConfigLoader::getInstance()->getThePatientInfoPatten();
     unsigned int index = 0;
@@ -261,6 +264,20 @@ void DAO_Sqlite::appendARow_Patient(const std::map<unsigned int,std::pair<QStrin
         if(it->second.primaryKey){
             auto it_find = patientInfos.find(it->first);
             if(it_find != patientInfos.end()){
+
+                this->getRowValueByItemValue_Patient(it->second.infoName,it_find->second.second,list);
+
+                if(list.size() == (int)patientInfos.size()){
+                    for(auto it_rep = patientInfos.begin();it_rep != patientInfos.end();it_rep++){
+                        if(list.at(it_rep->first) != it_rep->second.second){
+                            flag = false;
+                            break;
+                        }
+                    }
+                }
+
+                if(flag) return;
+
                 query.exec(QString("DELETE FROM %1 WHERE %2 = '%3';")
                                .arg(patientInfo_TableName)
                                .arg(it->second.infoName)
@@ -529,5 +546,20 @@ void DAO_Sqlite::getAllValueByKey_Patient(const QString & key,QStringList & resu
         if(query.value(0).isValid()){
             result.push_back(query.value(0).toString());
         }
+    }
+}
+
+void DAO_Sqlite::getRowValueByItemValue_Patient(const QString & key,const QString & value,QStringList & result) const{
+    QSqlQuery query(*this->theDataBase);
+    unsigned int size = 0;
+    query.exec(QString("select * from %1 where %2 = '%3';").arg(patientInfo_TableName).arg(key).arg(value));
+
+    qDebug()<<query.lastError();
+    size = query.record().count();
+    while(query.next()){
+        for(unsigned int i = 0;i<size;i++){
+            result.push_back(query.record().value(i).toString());
+        }
+        break;
     }
 }
